requirements.txt
fastapi
pydantic
uvicorn
jinja2
python-multipart  # for the Form
python-jose
sqlalchemy
pysqlite3 